+++ src/app/components/AddTodo/index.tsx
@@ -1,57 +1,19 @@
 import React, { useState } from "react";
-import styled from "styled-components";
-import { Button, Text, View, TextInput, Dimensions } from "react-native";
 import { postTodo } from "../../apiClient";
+import { Presentation, Props } from "./presentation";
 
-const AddTodo = () => {
+export default () => {
   const [todoTitle, setTodoTitle] = useState("");
-  const onPressSubmit = async () => {
+
+  const submit = async () => {
     await postTodo({ title: todoTitle });
     setTodoTitle("");
   };
-  return (
-    <Section>
-      <InlineInput>
-        <Label>やること</Label>
-        <TodoTitle
-          autoFocus
-          value={todoTitle}
-          onChangeText={setTodoTitle}
-          placeholder="ここに何か書いてね"
-        />
-      </InlineInput>
-      <ButtonArea>
-        <Button onPress={onPressSubmit} title="追加する" />
-      </ButtonArea>
-    </Section>
-  );
+
+  const props: Props = {
+    todoTitle,
+    onChangeTitle: setTodoTitle,
+    onSubmit: submit
+  };
+  return <Presentation {...props} />;
 };
-
-export default AddTodo;
-
-const Section = styled(View)`
-  flex: 1;
-  align-items: center;
-`;
-
-const InlineInput = styled(View)`
-  width: ${Dimensions.get("window").width};
-  display: flex;
-  background: #fff;
-  flex-direction: row;
-  align-items: center;
-  padding-left: 8;
-  padding-right: 8;
-`;
-
-const Label = styled(Text)`
-  width: 96;
-`;
-
-const TodoTitle = styled(TextInput)`
-  height: 48;
-`;
-
-const ButtonArea = styled(View)`
-  margin-top: 16;
-`;
